# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Student Profile Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: students
    description: Student profile operations
  - name: leaves
    description: Leave management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login to account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset link sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List students
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, graduated, suspended]
        - name: major
          in: query
          schema:
            type: string
        - name: year
          in: query
          schema:
            type: string
            enum: [freshman, sophomore, junior, senior, graduate]
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  /students/stats:
    get:
      operationId: getStudentStats
      tags: [students]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Student stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentStats"

  /students/recent:
    get:
      operationId: getRecentStudents
      tags: [students]
      summary: Get recently added students
      responses:
        "200":
          description: Recent students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"

  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /leave-types:
    get:
      operationId: listLeaveTypes
      tags: [leaves]
      summary: List available leave types
      responses:
        "200":
          description: Leave types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveType"

  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Leave applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveApplication"
    post:
      operationId: applyForLeave
      tags: [leaves]
      summary: Apply for leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Leave application created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveApplication"

  /leaves/stats:
    get:
      operationId: getLeaveStats
      tags: [leaves]
      summary: Get leave statistics
      responses:
        "200":
          description: Leave stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveStats"

  /leaves/{id}:
    patch:
      operationId: updateLeaveStatus
      tags: [leaves]
      summary: Update leave application status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveStatusUpdate"
      responses:
        "200":
          description: Updated leave application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveApplication"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    AuthUser:
      type: object
      required:
        - id
        - name
        - email
        - role
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, student, faculty]
        createdAt:
          type: string

    SignupInput:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [admin, student, faculty]

    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string

    ForgotPasswordInput:
      type: object
      required:
        - email
      properties:
        email:
          type: string

    Student:
      type: object
      required:
        - id
        - firstName
        - lastName
        - email
        - status
        - year
        - major
        - enrollmentDate
        - createdAt
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        enrollmentDate:
          type: string
        graduationYear:
          type: ["integer", "null"]
        major:
          type: string
        gpa:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, inactive, graduated, suspended]
        year:
          type: string
          enum: [freshman, sophomore, junior, senior, graduate]
        photoUrl:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string

    StudentInput:
      type: object
      required:
        - firstName
        - lastName
        - email
        - major
        - year
        - enrollmentDate
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        enrollmentDate:
          type: string
        graduationYear:
          type: integer
        major:
          type: string
        gpa:
          type: number
        status:
          type: string
          enum: [active, inactive, graduated, suspended]
        year:
          type: string
          enum: [freshman, sophomore, junior, senior, graduate]
        photoUrl:
          type: string
        address:
          type: string
        bio:
          type: string

    StudentUpdate:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
        phone:
          type: string
        dateOfBirth:
          type: string
        enrollmentDate:
          type: string
        graduationYear:
          type: integer
        major:
          type: string
        gpa:
          type: number
        status:
          type: string
          enum: [active, inactive, graduated, suspended]
        year:
          type: string
          enum: [freshman, sophomore, junior, senior, graduate]
        photoUrl:
          type: string
        address:
          type: string
        bio:
          type: string

    StudentStats:
      type: object
      required:
        - total
        - byStatus
        - byYear
        - byMajor
        - averageGpa
        - gpaDistribution
      properties:
        total:
          type: integer
        byStatus:
          type: object
          properties:
            active:
              type: integer
            inactive:
              type: integer
            graduated:
              type: integer
            suspended:
              type: integer
          required: [active, inactive, graduated, suspended]
        byYear:
          type: object
          properties:
            freshman:
              type: integer
            sophomore:
              type: integer
            junior:
              type: integer
            senior:
              type: integer
            graduate:
              type: integer
          required: [freshman, sophomore, junior, senior, graduate]
        byMajor:
          type: array
          items:
            type: object
            required: [major, count]
            properties:
              major:
                type: string
              count:
                type: integer
        averageGpa:
          type: ["number", "null"]
        gpaDistribution:
          type: array
          items:
            type: object
            required: [range, count]
            properties:
              range:
                type: string
              count:
                type: integer

    LeaveType:
      type: object
      required:
        - id
        - name
        - totalDays
      properties:
        id:
          type: integer
        name:
          type: string
        totalDays:
          type: integer

    LeaveApplication:
      type: object
      required:
        - id
        - applicantName
        - applicantEmail
        - leaveTypeName
        - fromDate
        - toDate
        - totalDays
        - reason
        - status
        - createdAt
      properties:
        id:
          type: integer
        applicantName:
          type: string
        applicantEmail:
          type: string
        leaveTypeName:
          type: string
        leaveTypeId:
          type: integer
        fromDate:
          type: string
        toDate:
          type: string
        totalDays:
          type: integer
        reason:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string

    LeaveInput:
      type: object
      required:
        - applicantName
        - applicantEmail
        - leaveTypeId
        - fromDate
        - toDate
        - totalDays
        - reason
      properties:
        applicantName:
          type: string
        applicantEmail:
          type: string
        leaveTypeId:
          type: integer
        fromDate:
          type: string
        toDate:
          type: string
        totalDays:
          type: integer
        reason:
          type: string

    LeaveStatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]

    LeaveStats:
      type: object
      required:
        - total
        - pending
        - approved
        - rejected
        - byType
      properties:
        total:
          type: integer
        pending:
          type: integer
        approved:
          type: integer
        rejected:
          type: integer
        byType:
          type: array
          items:
            type: object
            required: [name, count]
            properties:
              name:
                type: string
              count:
                type: integer