# lib/api-zod/src/generated/types/leaveInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Profile Management API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveInput {
  applicantName: string;
  applicantEmail: string;
  leaveTypeId: number;
  fromDate: string;
  toDate: string;
  totalDays: number;
  reason: string;
}