# lib/api-zod/src/generated/types/leaveStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Profile Management API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaveStatsByTypeItem } from './leaveStatsByTypeItem';

export interface LeaveStats {
  total: number;
  pending: number;
  approved: number;
  rejected: number;
  byType: LeaveStatsByTypeItem[];
}