# lib/api-zod/src/generated/types/leaveStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Profile Management API
 * OpenAPI spec version: 0.1.0
 */
import type { LeaveStatusUpdateStatus } from './leaveStatusUpdateStatus';

export interface LeaveStatusUpdate {
  status: LeaveStatusUpdateStatus;
}