# lib/api-zod/src/generated/types/studentStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Student Profile Management API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentStatsByMajorItem } from './studentStatsByMajorItem';
import type { StudentStatsByStatus } from './studentStatsByStatus';
import type { StudentStatsByYear } from './studentStatsByYear';
import type { StudentStatsGpaDistributionItem } from './studentStatsGpaDistributionItem';

export interface StudentStats {
  total: number;
  byStatus: StudentStatsByStatus;
  byYear: StudentStatsByYear;
  byMajor: StudentStatsByMajorItem[];
  /** @nullable */
  averageGpa: number | null;
  gpaDistribution: StudentStatsGpaDistributionItem[];
}